+++ config/config.go
@@ -129,6 +129,8 @@ const (
 
 	/* Kubernetes operator constants */
 	OperatorContainerName = "trident-operator"
+
+	DefaultAutosupportImage = "netapp/trident-autosupport:20.07.0"
 )
 
 var (